backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.26.0
pydantic==2.6.1
google-cloud-firestore==2.19.0
firebase-admin==6.4.0
